Create Main.c
This was an easy problem but with a strict time limit. Tried to solve it in ruby but the time limit exceeded everytime. But when the same logic is implemented in c then it works. This is because since there are a large number of test cases, the execution time heavily depends on the speed of getting the numbers from the console and printing it.
Will keep on working on this problem using ruby and implementing other sorting algorithms.